Generalize profile writing.
